fix(folders): move design system elements to ui folder
