--- src/sections/shared/Navbar.scss
@@ -1,9 +1,0 @@
-@import "~bootstrap/scss/functions";
-@import "~bootstrap/scss/variables";
-
-$navbar-light-brand-color: var(--bs-dataverse-orange);
-
-@import "~bootstrap/scss/mixins";
-@import "~bootstrap/scss/nav";
-@import "~bootstrap/scss/navbar";
-@import "~bootstrap/scss/containers";
